Extract isFullyOpen and theirAddress helpers
